2305-SCW-TUD: remove test jupyter notebooks from dry run of Python
It all goes as expected once you make sure that the autosave of Jupyterlab can be customised.
Its default value is 120 seconds and should be lowered to 2-5 seconds for gitautopush to catch
meaningful changes promptly.

The easist way via the Settings did not work for me. Instead I found the following tip
effective: https://www.webucator.com/article/change-default-autosave-interval-in-jupyterlab/