master/README: a new edition branch should have YYMM-WHAT-WHEN as source
This repo contains a branch for each workshop edition, which then stands
out in the list of branches for the name pattern starting with the year
and month. Clearly, each workshop edition produces a long-lived and
independent branch, like an octopus' tentacles.

The branch YYMM-WHAT-WHEN contains only a conveniently drafted README
branch. template for a new edition branch. It is then sensible that each
new branch developes from this clean support branch: a source branch is
input that GitHub asks indeed. The name YYMM-WHAT-WHEN also suggests the
recommended name pattern for the new edition branches, e.g.,
2302-SWC-TUD.