2310-gicodev-fixes: edit explanations of GitHub connections with SSH
I have commented on the early steps that participants should take to get
SSH going on Git Bash, that is, run it as local administrator. As @manuGil
suggested, this has been proven to forestall a category of issues down the
line and worked for the 2310-gitcodev editions.

Added a short explanation of what a positive `ssh -T <upstream address>`
is, so as to invite participants to recognize success or failure with a
little guidance