2310-gitcodev-fixes, Lesson3: cosmetic/formatting changes
Removed the markdown coding for shell/bash command, since indenting is
sufficient for a nice styling on GitHub and syntax highlighting is
unnecessary here. Only up to Section A.1.

Patterned paragraphs of Section A.1 as nested lists. This should make it
easier to recognize the step and substeps of the SSH-related procedure.

All document: removed trailing space, with no immediate benefit in sight.