2311-SWC: customize README for this workshop
At this moment it is not clear how many instructors want to use this branch to deposit their
instructor notes. (I will use it for Git.) Therefore, the istructions in the text are kept loose. Later it will perhaps
be possible to be more definite about the coverage of the lesson notes.