2311-SWC-TUD/README: specify that setup section is for instructors
The learners do read this page and may be confused by additional setup instructions.
Incidentally, the instructions for instructors are explained in a more elaborate fashion in the
master branch
https://github.com/4TUResearchData-Carpentries/workshop_notes/tree/master#3-preparations-for-coordinatorsinstructors.
The latter section may well need some more editing to become more turn-key that it has been so
far; if that is the case, every little editing helps.