2311-SWC-TUD/git.log: revise lesson note after the workshop
I have removed duplicates and commands resulting from errors and
explanations. I have not replayed the commands in the list, so be aware of
gaps and jumps. Added some comments using after the token # (users having
zsh instead of bash as the shell interpreter will not get this token
recognized).