.gitignore: enable tracking of log file
The .gitignore present before the Git Lesson seems an overkill with many
extensions not tracked to serve very special purposes. In particular, *.log
files were not tracked (under the heading of Django-oriented lines) while I
used it as an extension for the lesson notes. Not knowing of this long
.gitignore made my scripts not behave as expected, with associated debug
time find out that this .gitignore was the issue